Revert to earlier Tasks UI iteration
