+++ ui/v2.5/src/components/Settings/Tasks/ScanOptions.tsx
