+++ ui/v2.5/src/components/Settings/styles.scss
@@ -162,28 +162,20 @@
   }
 }
 
-.card.task-group {
-  padding-bottom: 0.5rem;
+.task-group {
   padding-top: 0.5rem;
 
-  .task {
+  &:not(:last-child) {
     padding-bottom: 0.5rem;
-    padding-top: 0.5rem;
+  }
+
+  .task {
+    &:not(:first-child) {
+      padding-top: 0.5rem;
+    }
 
     &:not(:last-child) {
-      border-bottom: 1px solid $dark-gray2;
+      padding-bottom: 1rem;
     }
-  }
-}
-
-.ellipsis-button {
-  .btn:first-child {
-    border-right: 1px solid $card-bg;
-  }
-
-  .btn:last-child {
-    border-left: 1px solid $card-bg;
-    padding-left: 0.25rem;
-    padding-right: 0.25rem;
   }
 }
